pyproject.toml
[project]
name = "fastapi-ddd"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aerich>=0.9.2",
    "aiomysql>=0.3.2",
    "asyncpg>=0.30.0",
    "cryptography>=46.0.3",
    "fastapi>=0.121.1",
    "pydantic-settings>=2.11.0",
    "python-dotenv>=1.2.1",
    "tomlkit>=0.13.3",
    "tortoise-orm>=0.25.1",
    "uvicorn>=0.38.0",
]

[tool.aerich]
tortoise_orm = "config.setting.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
